requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.12.0
postgres==4.0
psycopg2-binary>=2.9.10
sentence-transformers>=3.4.1,<4.0.0
torch>=2.9.0,<2.12.0
numpy>=2.2.0,<2.4.0
pydantic>=2.10.0
requests==2.32.3
tqdm==4.67.1
transformers>=4.56.0,<5.0.0
tokenizers>=0.21.0
python-multipart==0.0.6
scikit-learn>=1.6.0,<1.8.0
scipy>=1.15.0,<1.17.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
ollama==0.1.6
asyncpg>=0.30.0
pymupdf>=1.25.0  # Fitz (PyMuPDF) 로더
pdfplumber==0.11.5  # PDFPlumber 로더
unstructured==0.11.8  # UnstructuredPDFLoader
unstructured.pytesseract==0.3.15
pypdf==5.4.0  # PyPDFLoader
pytesseract>=0.3.13
pdf2image==1.17.0
Pillow>=10.4.0
opencv-python==4.11.0.86
docx2txt==0.9
markdown==3.5.2  # 마크다운 변환을 위한 패키지
openpyxl>=3.1.5
pandas>=2.2.0
nltk>=3.9.1